refactor : codeKata 관련 코드 수정
